chore: Move BackToHomeButton to components dir
